comparison/model_segm_comparison/cellpose/requirements_cellpose.txt
cellpose==3.0.5
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.4
cycler==0.12.1
fastremap==1.17.7
filelock==3.20.3
fonttools==4.61.1
fsspec==2026.1.0
google-api-core==2.29.0
google-auth==2.48.0
google-cloud-core==2.5.0
google-cloud-storage==3.9.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
idna==3.11
imagecodecs==2026.1.14
Jinja2==3.1.6
kiwisolver==1.4.9
llvmlite==0.46.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
natsort==8.4.0
networkx==3.6.1
numba==0.63.1
numpy==2.3.5
opencv-python-headless==4.13.0.90
packaging @ file:///C:/miniconda3/conda-bld/packaging_1761049096285/work
pandas==3.0.0
pillow==12.1.0
proto-plus==1.27.1
protobuf==6.33.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycocotools==2.0.11
pycparser==3.0
Pygments==2.19.2
pyparsing==3.3.2
PyQt6==6.10.2
PyQt6-Qt6==6.10.2
PyQt6_sip==13.11.0
pyqtgraph==0.14.0
python-dateutil==2.9.0.post0
QtPy==2.4.3
requests==2.32.5
roifile==2026.1.29
rsa==4.9.1
scipy==1.17.0
seaborn==0.13.2
six==1.17.0
statannotations==0.7.2
superqt==0.7.8
sympy==1.14.0
tifffile==2026.1.28
torch==2.10.0
tqdm==4.67.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
